divide & move files
