--- Engine/Physics/ColliderNode.cs
@@ -1,273 +1,0 @@
-using System;
-
-namespace Altseed2
-{
-    /// <summary>
-    /// コライダを管理するノード
-    /// </summary>
-    [Serializable]
-    public abstract class ColliderNode : TransformNode
-    {
-        public override Matrix44F AbsoluteTransform => _absoluteTransform;
-        private Matrix44F _absoluteTransform;
-
-        /// <summary>
-        /// コライダを取得する
-        /// </summary>
-        internal abstract Collider Collider { get; }
-
-        /// <summary>
-        /// <see cref="ColliderNode"/>の新しいインスタンスを生成する
-        /// </summary>
-        protected ColliderNode() { }
-
-        private static CollisionManagerNode SearchManagerFromChildren(Node node)
-        {
-            if (node == null) return null;
-            foreach (var current in node.Children)
-                if (current is CollisionManagerNode m)
-                    return m;
-            return null;
-        }
-
-        internal override void Added(Node owner)
-        {
-            base.Added(owner);
-            SearchManagerFromChildren(owner.Parent)?.AddCollider(this);
-        }
-
-        internal override void Removed()
-        {
-            SearchManagerFromChildren(Parent.Parent)?.RemoveCollider(this);
-            base.Removed();
-        }
-
-        internal abstract void UpdateCollider();
-
-        internal override void UpdateInheritedTransform()
-        {
-            _absoluteTransform = CalcInheritedTransform();
-        }
-    }
-
-    /// <summary>
-    /// 円形コライダを管理するノード
-    /// </summary>
-    [Serializable]
-    public class CircleColliderNode : ColliderNode
-    {
-        /// <summary>
-        /// 拡大率を適用する方法を表す
-        /// </summary>
-        [Serializable]
-        public enum ScaleCalcType
-        {
-            /// <summary>
-            /// Scale.Xで計算する
-            /// </summary>
-            X,
-            /// <summary>
-            /// Scale.Yで計算する
-            /// </summary>
-            Y,
-            /// <summary>
-            /// Scale.Lengthで計算する
-            /// </summary>
-            Length,
-            /// <summary>
-            /// X，Yの最小値で計算する
-            /// </summary>
-            Min,
-            /// <summary>
-            /// X，Yの最大値で計算する
-            /// </summary>
-            Max
-        }
-
-        /// <summary>
-        /// 使用するコライダを取得する
-        /// </summary>
-        public CircleCollider CircleCollider { get; }
-        internal override Collider Collider => CircleCollider;
-
-        /// <summary>
-        /// 衝突半径を取得または設定する
-        /// </summary>
-        public float Radius
-        {
-            get => _radius;
-            set
-            {
-                if (_radius == value) return;
-                _radius = value;
-                AdjustSize();
-            }
-        }
-        private float _radius;
-
-        /// <summary>
-        /// 拡大率の計算方法を取得または設定します。
-        /// </summary>
-        /// <remarks>既定値：<see cref="ScaleCalcType.Max"/></remarks>
-        public ScaleCalcType ScaleType { get; set; } = ScaleCalcType.Max;
-
-        public override Vector2F Size
-        {
-            get => base.Size;
-            set
-            {
-                if (base.Size == value) return;
-                base.Size = value;
-
-                if (Radius != 0) Scale = value / Radius / 2f;
-            }
-        }
-        /// <summary>
-        /// 既定の<see cref="Altseed2.CircleCollider"/>を使用して<see cref="CircleColliderNode"/>の新しいインスタンスを生成する
-        /// </summary>
-        public CircleColliderNode() : this(new CircleCollider()) { }
-
-        /// <summary>
-        /// 指定した<see cref="Altseed2.CircleCollider"/>を使用して<see cref="CircleColliderNode"/>の新しいインスタンスを生成する
-        /// </summary>
-        /// <param name="collider">使用する<see cref="Altseed2.CircleCollider"/>のインスタンス</param>
-        /// <exception cref="ArgumentNullException"><paramref name="collider"/>がnull</exception>
-        public CircleColliderNode(CircleCollider collider)
-        {
-            CircleCollider = collider ?? throw new ArgumentNullException(nameof(collider), "引数がnullです");
-        }
-
-        public override void AdjustSize()
-        {
-            Size = Scale * Radius * 2;
-        }
-
-        internal override void UpdateCollider()
-        {
-            UpdateInheritedTransform();
-
-            MathHelper.CalcFromTransform(AbsoluteTransform, out var position, out var scale, out var angle);
-            Collider.Position = position - new Vector2F(Radius, Radius);
-            Collider.Rotation = MathHelper.DegreeToRadian(angle);
-
-            CircleCollider.Radius = Radius * (ScaleType switch
-            {
-                ScaleCalcType.X => scale.X,
-                ScaleCalcType.Y => scale.Y,
-                ScaleCalcType.Length => scale.Length,
-                ScaleCalcType.Min => Math.Min(scale.X, scale.Y),
-                ScaleCalcType.Max => Math.Max(scale.X, scale.Y),
-                _ => 1.0f
-            });
-        }
-    }
-
-    /// <summary>
-    /// ポリゴンコライダを管理するノード
-    /// </summary>
-    [Serializable]
-    public class PolygonColliderNode : ColliderNode
-    {
-        /// <summary>
-        /// 使用するコライダを取得する
-        /// </summary>
-        public PolygonCollider PolygonCollider { get; }
-        internal override Collider Collider => PolygonCollider;
-
-        /// <summary>
-        /// 頂点情報の配列を取得または設定する
-        /// </summary>
-        /// <exception cref="ArgumentNullException">設定しようとした値がnull</exception>
-        public Vector2F[] Vertexes
-        {
-            get => _vertexes;
-            set
-            {
-                _vertexes = value ?? throw new ArgumentNullException(nameof(value), "引数がnullです");
-                AdjustSize();
-            }
-        }
-        private Vector2F[] _vertexes = Array.Empty<Vector2F>();
-
-        /// <summary>
-        /// コンテンツのサイズを取得します。
-        /// </summary>
-        public new Vector2F Size => base.Size;
-
-        /// <summary>
-        /// 既定の<see cref="Altseed2.PolygonCollider"/>を使用して<see cref="PolygonColliderNode"/>の新しいインスタンスを生成する
-        /// </summary>
-        public PolygonColliderNode() : this(new PolygonCollider()) { }
-
-        /// <summary>
-        /// 指定した<see cref="Altseed2.PolygonCollider"/>を使用して<see cref="PolygonColliderNode"/>の新しいインスタンスを生成する
-        /// </summary>
-        /// <param name="collider">使用する<see cref="Altseed2.PolygonCollider"/>のインスタンス</param>
-        /// <exception cref="ArgumentNullException"><paramref name="collider"/>がnull</exception>
-        public PolygonColliderNode(PolygonCollider collider)
-        {
-            PolygonCollider = collider ?? throw new ArgumentNullException(nameof(collider), "引数がnullです");
-        }
-
-        public override void AdjustSize()
-        {
-            MathHelper.GetMinMax(out var min, out var max, _vertexes);
-            base.Size = max - min;
-        }
-
-        internal override void UpdateCollider()
-        {
-            UpdateInheritedTransform();
-
-            MathHelper.CalcFromTransform(AbsoluteTransform, out var position, out var scale, out var angle);
-            Collider.Position = position;
-            Collider.Rotation = MathHelper.DegreeToRadian(angle);
-
-            var count = _vertexes.Length;
-            var array = new Vector2F[count];
-            for (int i = 0; i < count; i++) array[i] = _vertexes[i] * Scale;
-
-            PolygonCollider.VertexArray = _vertexes;
-        }
-    }
-
-    /// <summary>
-    /// 短形コライダを管理するノード
-    /// </summary>
-    [Serializable]
-    public class RectangleColliderNode : ColliderNode
-    {
-        /// <summary>
-        /// 使用するコライダを取得する
-        /// </summary>
-        public RectangleCollider RectangleCollider { get; }
-        internal override Collider Collider => RectangleCollider;
-
-        /// <summary>
-        /// 既定の<see cref="Altseed2.RectangleCollider"/>を使用して<see cref="RectangleColliderNode"/>の新しいインスタンスを生成する
-        /// </summary>
-        public RectangleColliderNode() : this(new RectangleCollider()) { }
-
-        /// <summary>
-        /// 指定した<see cref="Altseed2.RectangleCollider"/>を使用して<see cref="RectangleColliderNode"/>の新しいインスタンスを生成する
-        /// </summary>
-        /// <param name="collider">使用する<see cref="Altseed2.RectangleCollider"/>のインスタンス</param>
-        /// <exception cref="ArgumentNullException"><paramref name="collider"/>がnull</exception>
-        public RectangleColliderNode(RectangleCollider collider)
-        {
-            RectangleCollider = collider ?? throw new ArgumentNullException(nameof(collider), "引数がnullです");
-        }
-
-        public override void AdjustSize() { }
-
-        internal override void UpdateCollider()
-        {
-            UpdateInheritedTransform();
-
-            MathHelper.CalcFromTransform(AbsoluteTransform, out var position, out var scale, out var angle);
-            Collider.Position = position;
-            Collider.Rotation = MathHelper.DegreeToRadian(angle);
-            RectangleCollider.Size = Size * Scale;
-        }
-    }
-}
